aws-sdk-lambda: add unit test for lambda
The function invocation methods offered by the aws-sdk-lambda gem take
the `:function_name` argument in the form of `<function name>:<alias>`
or `<function name>:<version>` when an alias or version is involved. The
agent will simply preserve the extra info and the colon delimiter and
segment naming and AWS arn creation will simply end up doing the right
thing with the extra alias or version info.