aws-sdk-lambda: improve error handling
use separate error handling for errors raised by the gem itself and
errors raised by the agent.

Errors raised by the gem need to be noticed by the agent (sent to New
Relic) and re-raised to the user's app.

Errors raised by the agent need to be logged but not impact the
returning of the successful gem invocation response.

Co-authored-by: Kayla Reopelle <87386821+kaylareopelle@users.noreply.github.com>