aws-sdk-lambda fix
commit 7b44eb0d249917ae26c3439cf6f722a668e86486 sought to improve error
handling, but incorrectly misused an `ensure` block to attempt to always
return the raw client response.

this faulty logic has been fixed. internal agent response handling
errors are logged as desired, not re-raised to the user's app, and the
user's app is guaranteed to always get the raw original response.