Add agent control health checks
When the agent recognizes it is running in an agent control
environment, it will start automatic health checks that will
create a new file at a configured destination at a given
frequency that provides details about the last reported status
of the agent.

When the agent is not seen within an agent control environment,
files will not be created.