Remove deprecated_disabled_configured
We removed all the disabled_* configuration options that correlate
to an instrumentation.* configuration option in our last major version.
The disabled_* options
that remain will not be replaced with instrumentation.* options because
they do not use alias method chaining or module prepending or are not
related to instrumentation-related, but rather specific features.

Also, add alias for disable_active_job.